Adding Simple Gin Unit Tests (#24)
Add a very basic gin test example where we
get the index page and check it contains
a particular post in the mock database.

To complete this, I had to mock the database
by turning the previous declaration into
an interface, and moving the concrete types
into `SqlDatabase`. The mock is done in `DatabaseMock`.