System Tests with full in memory DB support (#57)
This PR implements full system tests / end2end tests using an in memory
mysql databse. The motivation for this work was to get coverage for the
lines that require a real db connection.

These tests are still expensive to run, as the internal DB server needs
to be started (~50ms), so only need to have system tests for fully
testing a feature. Leave smaller tests for unit tests.